+++ src/Resources/SernosNoteQueryResource.cs
@@ -2,10 +2,6 @@
 {
     public class SernosNoteQueryResource
     {
-        public string SernosGroup { get; set; }
-
         public int SernosNumber { get; set; }
-
-        public string TransCode { get; set; }
     }
 }
