+++ src/Service.Host/client/src/reducers/__tests__/itemStoreFactory.js
@@ -28,6 +28,7 @@ describe('item store reducer factory', () => {
         };
 
         const expected = {
+            item: {},
             loading: true,
             editStatus: 'view'
         };
