+++ src/Service.Host/client/src/reducers/__tests__/sernosConfig.js
@@ -14,6 +14,7 @@ describe('sernos config reducer', () => {
         };
 
         const expected = {
+            item: {},
             loading: true,
             editStatus: 'view'
         };
